requirements-rag.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
chromadb>=0.4.22
google-cloud-storage>=2.13.0
google-generativeai>=0.3.2
langchain>=0.1.0
langchain-google-genai>=0.0.5
sentence-transformers>=2.2.2
pandas>=2.0.0
